Don't need to set slot and videoSlot
This is handled by the VPAIDHTML5Client lib. It sets the videoSlot for us. And it also
uses the correct slot inside the iframe it creates. This fixes a 'click not working'
issue with some VPAID creatives. And it will also solve the width/height problem - when
the creative doesn't set the video element's width/height resulting in larger than
desired video..  because the video is not inside iframe where the slot is.

Need to revisit `videoInstance` option and its influence on slot and videoSlot. Also
need to consider exposing VPAIDHTML5Client's 'template' and 'extraOptions' config
options.